refactor: extract common hook for test settings
